New animation controller
A new controller which doesn't look AS bad as the first. It is much simpiler with no masks and a little less parameters. Removed old animations that were not being used. Added many new animations from a sword and shield set.

Tweeked camera collision to make it a bit better and created a third person camera collision testing room.

Added a singleton script to enforce some classes can only be instanciated once.